Send solr request only once [#19] (#20)
The solr request to the backend is sent 2 times. This fixes it and
only 1 is sent.

By the way, this triggered a bug in asyncconnect. There was a race
condition and sometimes the results will be ignored completely, that is
both of the two updates were ignored.

It's beyond the scope to fix asyncconnect, but sending only 1 request
instead of the double should fix this.

Co-authored-by: Balázs Reé <ree@greenfinity.hu>